Moved CycleSwticher component under Header (not FraHeader)
