+++ src/client/components/Navigation/NavAssessment/Header/Title/Title.tsx
@@ -7,9 +7,8 @@ import { useAssessment } from '@client/store/assessment'
 import { DataLockActions, useIsDataLocked } from '@client/store/ui/dataLock'
 import { useIsDataExportView } from '@client/hooks'
 import Icon from '@client/components/Icon'
+import CycleSwitcher from '@client/components/PageLayout/Header/CycleSwitcher'
 import { Breakpoints } from '@client/utils'
-
-import CycleSwitcher from './CycleSwitcher'
 
 type Props = {
   lockEnabled: boolean
@@ -31,9 +30,9 @@ const Title: React.FC<Props> = (props) => {
   return (
     <div className="nav-assessment-header__lock">
       <div>
-        {i18n.t(`assessment.${type}`)} <CycleSwitcher />
+        {i18n.t<string>(`assessment.${type}`)} <CycleSwitcher />
         {isDataExportView ? ` - ${i18n.t('common.dataExport')}` : ''}
-        {deskStudy && <div className="desk-study">({i18n.t('assessment.deskStudy')})</div>}
+        {deskStudy && <div className="desk-study">({i18n.t<string>('assessment.deskStudy')})</div>}
       </div>
 
       {lockEnabled && (
