+++ src/client/components/PageLayout/Header/CycleSwitcher/CycleSwitcher.scss
