+++ src/client/components/PageLayout/Header/CycleSwitcher/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CycleSwitcher'
